Fix up import cycles
