+++ abstract_ranker/llm_utils.py
@@ -1,30 +1,8 @@
 from typing import Any, Callable, Dict, List
 
-from pydantic import BaseModel, Field
-
+from abstract_ranker.data_model import AbstractLLMResponse
 from abstract_ranker.local_llms import query_hugging_face
 from abstract_ranker.openai_utils import query_gpt
-
-
-class AbstractLLMResponse(BaseModel):
-    "Result back from LLM grading of an abstract"
-
-    summary: str = Field(..., title="The one-line summary of the abstract")
-    experiment: str = Field(
-        ...,
-        title="The most likely experiment associated with this work (ATLAS, CMS, LHCb, "
-        "MATHUSLA, etc.). Blank if unknown.",
-    )
-    keywords: List[str] = Field(
-        ..., title="List of keywords associated with the abstract"
-    )
-    interest: str = Field(
-        ..., title="Interest level in the abstract (high, medium, or low)"
-    )
-    explanation: str = Field(
-        ...,
-        title="Explanation of the interest level in the abstract",
-    )
 
 
 _llm_dispatch: Dict[str, Callable[[str, Dict[Any, Any]], AbstractLLMResponse]] = {
